add skill record decoding/encoding but should notice that the fourth job skills will cause 38 problem maybe caused by IsSkillNeedMasterLevel. start working on basic stat but still in poor progress.

+++ Database/GW_CharacterList.cpp
@@ -8,7 +8,7 @@ void GW_CharacterList::Load(int nAccountID, int nWorldID)
 	queryStatement.execute();
 
 	Poco::Data::RecordSet recordSet(queryStatement);
-	nCount = recordSet.rowCount();
+	nCount = (int)recordSet.rowCount();
 	aCharacterList.resize(nCount);
 	for (int i = 0; i < recordSet.rowCount(); ++i, recordSet.moveNext())
 	{
